It turns out that reopening the USB port is unimportant.
Since it causes the device to spit out EPIPE when using pyusb (but not when using ugen??)
just get rid of it entirely.
What I *thought* was a necessary feature was actually just adding time,
and tweaking sleep() can do that better.
TODO: fix my patch to swetland's original version in light of this.